mssql: enable _insertID() again
In ADOdb 5.21.0, for performance reasons [1] the _query() method no
longer retrieves the last inserted Id.

This introduced a regression, as it was only possible to retrieve the
Insert Id for "normal" queries, but not parameterized ones as
Insert_ID() always returned null in this case.

To maintain the driver's performance, we now allow the client to
decide whether insert queries should execute the extra SCOPE_IDENTITY()
call, by providing a new enableLastInsertID() method.

Fixes #692

[1]: https://github.com/ADOdb/ADOdb/issues/185#issuecomment-169057606